requirements.txt
streamlit
pandas
qrcode
opencv-python-headless
numpy
Pillow